html/semantics/embedded-content/media-elements/interfaces/TextTrack/language.html

<!doctype html>
<title>TextTrack.language</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
setup(function(){
    window.video = document.createElement('video');
    window.t1 = video.addTextTrack('subtitles', 'foo', 'foo');
    window.track = document.createElement('track');
    track.setAttribute('srclang', 'bar');
    video.appendChild(track);
    window.t2 = track.track;
});
test(function(){
    assert_equals(t1.language, 'foo');
    assert_equals(t2.language, 'bar');
    track.srclang = 'baz';
    assert_equals(t2.language, 'baz');
    track.removeAttribute('srclang');
    assert_equals(t2.language, '');
});
test(function(){
    track.srclang = 'zh-Hant-CN-x-private1-private2';
    assert_equals(t2.language, 'zh-Hant-CN-x-private1-private2', 'IDL attribute');
    track.setAttribute('srclang', 'i-klingon');
    assert_equals(t2.language, 'i-klingon', 'content attribute');
}, document.title+', BCP47 language tag');
</script>
